pyproject.toml
# Big thanks to https://medium.com/georgian-impact-blog/python-tooling-makes-a-project-tick-181d567eea44

[tool.coverage.run]
branch = true
source = ["hough"]
concurrency = ["multiprocessing"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "except KeyboardInterrupt",
    "if __name__ == .__main__.:",
]

[tool.flakehell]
max_line_length = 88
format = "colored"
exclude = [".venv",".git","__pycache__","doc/*","build","dist",".tox"]
select = ["C", "E", "F", "W", "B", "B950"]
ignore = ["E203", "E501", "W503"]
show_source = true


[tool.poetry]
name = "hough"
version = "0.2.0"
description = "Skew detection and correction in scanned images"
authors = ["qu1j0t3 <support@telegraphics.com.au>", "Joan Touzet <wohali@apache.org>"]
license = "GPL-2.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6.1"
docopt-ng = "^0.7.2"
imageio = "^2.8.0"
numpy = "^1.18.2"
scikit-image = "^0.16.2"
scipy = "^1.4.1"
termplotlib = "^0.2.4"
pip = "20.0.2"
pymupdf = "^1.16.15"
filetype = "^1.0.6"
tqdm = "^4.43.0"
coverage = {extras = ["toml"], version = "^5.0.4"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
#flake8 = "^3.7.9"
isort = "^4.3.21"
pytest = "^5.4"
pre-commit = "^2.2.0"
pytest-cov = "^2.8.1"
pytest-mock = "^2.0.0"
coverage = {extras = ["toml"], version = "^5.0.4"}
tox = "^3.14.6"
flakehell = "^0.3.3"

[tool.poetry.scripts]
hough = 'hough.cli:run'

[tool.isort]
known_third_party = ["docopt", "filetype", "fitz", "imageio", "numpy", "pytest", "scipy", "skimage", "termplotlib", "tqdm"]
multi_line_output = 3
lines_after_imports = 2
force_grid_wrap = 0
combine_as_imports = true
include_trailing_comma = true

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = true
envlist = py36,py37,py38

[testenv]
whitelist_externals = poetry
commands =
    poetry install -v
    poetry run pytest tests/ -v -x --cov=./ --cov-config=setup.cfg --cov-report=term --cov-report=html --cov-report=xml
"""

[tool.black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
